requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
anyio==4.3.0
    # via httpx
attrs==23.2.0
    # via jsonschema
betterproto==2.0.0b6
    # via envoy-data-plane
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charmed-kubeflow-chisme==0.3.0
    # via -r requirements.in
charset-normalizer==3.3.2
    # via requests
cosl==0.0.10
    # via -r requirements.in
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
envoy-data-plane==0.2.5
    # via -r requirements.in
exceptiongroup==1.2.0
    # via anyio
grpclib==0.4.7
    # via betterproto
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via grpclib
hpack==4.0.0
    # via h2
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via lightkube
hyperframe==6.0.1
    # via h2
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.3.2
    # via jsonschema
jinja2==3.1.3
    # via charmed-kubeflow-chisme
jsonschema==4.17.3
    # via serialized-data-interface
lightkube==0.15.2
    # via charmed-kubeflow-chisme
lightkube-models==1.29.0.7
    # via lightkube
markupsafe==2.1.5
    # via jinja2
multidict==6.0.5
    # via grpclib
oci-image==1.0.0
    # via -r requirements.in
ops==2.11.0
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
    #   cosl
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
pkgutil-resolve-name==1.3.10
    # via jsonschema
pyrsistent==0.20.0
    # via jsonschema
python-dateutil==2.9.0.post0
    # via betterproto
pyyaml==6.0.1
    # via
    #   cosl
    #   lightkube
    #   ops
    #   serialized-data-interface
requests==2.31.0
    # via serialized-data-interface
ruamel-yaml==0.18.6
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
serialized-data-interface==0.7.0
    # via charmed-kubeflow-chisme
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tenacity==8.2.3
    # via charmed-kubeflow-chisme
typing-extensions==4.10.0
    # via
    #   anyio
    #   cosl
urllib3==2.2.1
    # via requests
websocket-client==1.7.0
    # via ops
zipp==3.18.1
    # via importlib-resources
